Prefer SSL if original URI or temp URI is HTTPS (#6)
Currently the REST proxy returns HTTP endpoints for the consumer unless it's configured to listen on HTTPS. However, some setups of REST proxy (like ours) need to run behind a load balancer or reverse proxy that handles SSL termination and forwards the request to the REST proxy instance over regular HTTP.

While this should be handled by the REST proxy itself, there currently is no mechanism for doing so. I will try to submit a PR to Confluent to fix this issue, but in the meantime here is a temporary fix that should be good for the majority of use cases.